app3/requirements.txt
agent-framework
python-dotenv>=1.0.0
azure-identity>=1.15.0
agent-framework-a2a